+++ db/migrations/0034_regular_dreaming_celestial.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "hce"."vacunaciones" ADD COLUMN "costo" numeric(12, 2);